Extract filetypes into its own type
